+++ src/main/js/types/PullRequest.ts
@@ -21,6 +21,7 @@ export type PullRequest = BasicPullRequest & {
   tasks: Tasks;
   sourceRevision: string;
   targetRevision: string;
+  markedAsReviewed: string[];
 };
 
 export type Location = {
@@ -34,6 +35,7 @@ export type BasicComment = {
   comment: string;
   id: string;
   type: string;
+  mentions: Mention[];
 };
 
 export type Comment = BasicComment & {
@@ -48,6 +50,12 @@ export type Comment = BasicComment & {
   context?: Context;
   _links: Links;
   _embedded?: { [key: string]: any };
+};
+
+export type Mention = {
+  id: string;
+  displayName: string;
+  mail: string;
 };
 
 export type Context = {