+++ src/test/java/com/cloudogu/scm/review/pullrequest/service/DefaultPullRequestServiceTest.java
@@ -40,6 +40,7 @@ import static com.cloudogu.scm.review.pullrequest.service.PullRequestRejectedEve
 import static com.cloudogu.scm.review.pullrequest.service.PullRequestStatus.MERGED;
 import static com.cloudogu.scm.review.pullrequest.service.PullRequestStatus.OPEN;
 import static com.cloudogu.scm.review.pullrequest.service.PullRequestStatus.REJECTED;
+import static com.google.common.collect.ImmutableSet.of;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
@@ -492,9 +493,60 @@ class DefaultPullRequestServiceTest {
 
       assertThat(eventCaptor.getAllValues()).isEmpty();
     }
+
+    @Test
+    void shouldMarkAsReviewed() {
+      service.markAsReviewed(REPOSITORY, pullRequest.getId(), "some/file", new User("user"));
+
+      verify(store).update(argThat(pr -> {
+        assertThat(pr.getReviewMarks())
+          .contains(new ReviewMark("some/file", "user"));
+        return true;
+      }));
+    }
+
+    @Test
+    void shouldMarkAsNotReviewed() {
+      pullRequest.setReviewMarks(of(new ReviewMark("some/file", "user")));
+
+      service.markAsNotReviewed(REPOSITORY, pullRequest.getId(), "some/file", new User("user"));
+
+      verify(store).update(argThat(pr -> {
+        assertThat(pr.getReviewMarks()).isEmpty();
+        return true;
+      }));
+    }
+
+    @Test
+    void shouldRemoveReviewMarks() {
+      pullRequest.setReviewMarks(of(
+        new ReviewMark("some/path", "dent"),
+        new ReviewMark("some/other/path", "dent")
+      ));
+
+      service.removeReviewMarks(REPOSITORY, pullRequest.getId(), of(new ReviewMark("some/path", "dent")));
+
+      verify(store).update(argThat(pr -> {
+        assertThat(pr.getReviewMarks())
+          .contains(new ReviewMark("some/other/path", "dent"));
+        return true;
+      }));
+    }
+
+    @Test
+    void shouldDoNothingWhenReviewMarksToBeRemovedAreEmpty() {
+      pullRequest.setReviewMarks(of(
+        new ReviewMark("some/path", "dent"),
+        new ReviewMark("some/other/path", "dent")
+      ));
+
+      service.removeReviewMarks(REPOSITORY, pullRequest.getId(), of());
+
+      verify(store, never()).update(any());
+    }
   }
 
   private PullRequest createPullRequest(String id, Instant creationDate, Instant lastModified) {
-    return new PullRequest(id, "source", "target", "pr", "description", null, creationDate, lastModified, OPEN, emptySet(), new HashMap<>(), "", "");
+    return new PullRequest(id, "source", "target", "pr", "description", null, creationDate, lastModified, OPEN, emptySet(), new HashMap<>(), "", "", emptySet());
   }
 }